+++ app/src/main/java/app/revanced/manager/ui/screen/PatcherScreen.kt
@@ -512,8 +512,6 @@ fun PatcherScreen(
                     Steps(
                         category = category,
                         steps = steps,
-                        stepCount = if (category == StepCategory.PATCHING) viewModel.patchesProgress else null,
-                        stepProgressProvider = viewModel,
                         isExpanded = if (autoCollapsePatcherSteps) {
                             expandedCategory == category
                         } else {